Separate legacy code to original files in case someone did a direct export
