+++ FrEIA/framework/__init__.py
@@ -4,6 +4,8 @@ inferring the order that the nodes have to be executed in forward and backward
 direction.
 """
 
+from .graph_inn import *
+from .sequence_inn import *
 from .reversible_graph_net import *
 from .reversible_sequential_net import *
